Bitcore query for all btc address transaction data displayed by the explorer UI
