+++ test/bitcore-db/queries.test.ts
@@ -11,8 +11,23 @@ import {
 } from '../../lib/bitcore-db/queries';
 
 test('can fetch address txs', async () => {
-  const tx = await getAddressTransactions('16iBt6f8ZhbutEE4sb1c2hZ8PHhVnabmv4', 0, 3);
-  console.log('ok');
+  const bitcoreTXs = await getAddressTransactions('16iBt6f8ZhbutEE4sb1c2hZ8PHhVnabmv4', 0, Number.MAX_SAFE_INTEGER);
+  const bitcoreTX = bitcoreTXs.find(t => t.txid === '7daa187ae06803d4def226503cbfee0b054dc4c4be07d203fa9c0445dd21d60d');
+
+  expect(bitcoreTX.action).toBe('sent');
+  expect(bitcoreTX.address).toBe('16iBt6f8ZhbutEE4sb1c2hZ8PHhVnabmv4');
+  expect(bitcoreTX.blockHash).toBe('00000000000000000007bdef8e74912573d72f1793509e003f8bcb0c14c2901a');
+  expect(bitcoreTX.blockHeight).toBe(606697);
+  expect(bitcoreTX.confirmations).toBeGreaterThan(100);
+  expect(bitcoreTX.date).toBe('2019-12-05T00:15:25.000Z');
+  expect(bitcoreTX.fee).toBe(1424);
+  expect(bitcoreTX.totalTransferred).toBe(395584);
+  expect(bitcoreTX.txid).toBe('7daa187ae06803d4def226503cbfee0b054dc4c4be07d203fa9c0445dd21d60d');
+  expect(bitcoreTX.outputs).toEqual([
+    { address: false, value: 0 },
+    { address: "16iBt6f8ZhbutEE4sb1c2hZ8PHhVnabmv4", value: 235584 },
+    { address: "1111111111111111111114oLvT2", value: 160000 }
+  ]);
 });
 
 test('can fetch address balance', async () => {
